Listing search improvements
* Added pagination to listing search
* Type of price fields in DTO, and price columns in db changed to double

+++ src/main/java/com/denmit99/hairbnb/model/bo/ListingBO.java
@@ -26,7 +26,7 @@ public class ListingBO {
 
     private String address;
 
-    private Integer pricePerNight;
+    private Double pricePerNight;
 
     private Currency currency;
