+++ src/main/java/com/denmit99/hairbnb/model/bo/ListingLightBO.java
@@ -21,7 +21,7 @@ public class ListingLightBO {
 
     private String address;
 
-    private Integer pricePerNight;
+    private Double pricePerNight;
 
     private Currency currency;
