+++ src/main/java/com/denmit99/hairbnb/model/dto/ListingDTO.java
@@ -24,7 +24,7 @@ public class ListingDTO {
 
     private String address;
 
-    private Integer pricePerNight;
+    private Double pricePerNight;
 
     private Currency currency;
